+++ hyde.py
@@ -1,12 +1,33 @@
-import json
-import langchain
-
+from langchain.prompts import PromptTemplate
+from langchain.chains import LLMChain
+from cat.mad_hatter.decorators import plugin
+from pydantic import BaseModel, Field
 from cat.log import log
 from cat.mad_hatter.decorators import hook
 
 
-with open("cat/plugins/ccat_hyde/settings.json", "r") as json_file:
-    settings = json.load(json_file)
+class MySettings(BaseModel):
+    hyde_prompt: str = Field(
+                title="HyDe prompt",
+                default="""You will be given a sentence.
+    If the sentence is a question, convert it to a plausible answer. If the sentence does not contain a question, 
+    just repeat the sentence as is without adding anything to it.
+
+    Examples:
+    - what furniture there is in my room? --> In my room there is a bed, a wardrobe and a desk with my computer
+    - where did you go today --> today I was at school
+    - I like ice cream --> I like ice cream
+    - how old is Jack --> Jack is 20 years old
+
+    Sentence:
+    - {input} -->""",
+                extra={"type": "TextArea"}
+        )
+
+@plugin
+def settings_schema():
+    return MySettings.schema()
+
 
 # Keys
 HYDE_ANSWER       = "hyde_answer"
@@ -16,22 +37,28 @@ AVERAGE_EMBEDDING = "average_embedding"
 @hook(priority=1)
 def cat_recall_query(user_message, cat):
 
+    #This method not work yet, but in the future I hope it will work
+    settings = cat.mad_hatter.get_plugin().load_settings() 
+    
+    log.debug(f" --------- ACQUIRE SETTINGS ---------")
+    log.debug(f"settings: {settings}")
+
     # Make a prompt from template
-    hypothesis_prompt = langchain.PromptTemplate(
+    hypothesis_prompt = PromptTemplate(
         input_variables=["input"],
         template=settings["hyde_prompt"]
     )
 
     # Run a LLM chain with the user message as input
-    hypothesis_chain = langchain.chains.LLMChain(prompt=hypothesis_prompt, llm=cat._llm)
+    hypothesis_chain = LLMChain(prompt=hypothesis_prompt, llm=cat._llm)
     answer = hypothesis_chain(user_message)
     
     # Save HyDE answer in working memory
     cat.working_memory[HYDE_ANSWER] = answer["text"]
     
-    print("------------- HYDE -------------")
-    print(f"user message: {user_message}")
-    print(f"hyde answer: {answer['text']}")
+    log.warning("------------- HYDE -------------")
+    log.warning(f"user message: {user_message}")
+    log.warning(f"hyde answer: {answer['text']}")
     
     return user_message
 
@@ -39,10 +66,10 @@ def cat_recall_query(user_message, cat):
 # Calculates the average between the user's message embedding and the Hyde response embedding
 def _calculate_vector_average(config: dict, cat):
     
-    # If average embedding not exists & hyde answer exists ..
-    if AVERAGE_EMBEDDING not in cat.working_memory.keys() and HYDE_ANSWER in cat.working_memory.keys():
+    # If hyde answer exists, calculate and set average embedding
+    if HYDE_ANSWER in cat.working_memory.keys():
         
-        # Get user message embedding
+       # Get user message embedding
         user_embedding = config['embedding']
         
         # Calculate hyde embedding from hyde answer
@@ -53,18 +80,23 @@ def _calculate_vector_average(config: dict, cat):
         average_embedding = [(x + y)/2 for x, y in zip(user_embedding, hyde_embedding)]
         cat.working_memory[AVERAGE_EMBEDDING] = average_embedding
 
-        #print(f"user_embedding:    {user_embedding}")
-        #print(f"hyde_embedding:    {hyde_embedding}")
-        #print(f"average_embedding: {average_embedding}")
+        log.debug(f" --------- CALCULATE AVERAGE ---------")
+        log.debug(f"hyde answer:       {hyde_answer}")
+        log.debug(f"user_embedding:    {user_embedding}")
+        log.debug(f"hyde_embedding:    {hyde_embedding}")
+        log.debug(f"average_embedding: {average_embedding}")
 
         # Delete Hyde Answer from working memory
         del cat.working_memory[HYDE_ANSWER]
-    
+
     # If average embedding exists, set the embedding
     if AVERAGE_EMBEDDING in cat.working_memory.keys():
         average_embedding = cat.working_memory[AVERAGE_EMBEDDING]
         config['embedding'] = average_embedding
-
+        
+        log.debug(f" --------- SET EMBEDDING ---------")
+        log.debug(f"average_embedding: {average_embedding}")
+        
 
 @hook(priority=1)
 def before_cat_recalls_episodic_memories(config: dict, cat):
