--- settings.json
@@ -1,12 +1,0 @@
-{
-    "title": "MySettings",
-    "type": "object",
-    "properties": {
-        "hyde_prompt": {
-            "title": "HyDe prompt",
-            "default": "You will be given a sentence.\n    If the sentence is a question, convert it to a plausible answer. If the sentence does not contain a question, \n    just repeat the sentence as is without adding anything to it.\n\n    Examples:\n    - what furniture there is in my room? --> In my room there is a bed, a wardrobe and a desk with my computer\n    - where did you go today --> today I was at school\n    - I like ice cream --> I like ice cream\n    - how old is Jack --> Jack is 20 years old\n\n    Sentence:\n    - {input} -->",
-            "extra": {"type": "TextArea"},
-            "type": "string"
-        }
-    }
-}
